+++ virustotal/src/index.ts
@@ -2,34 +2,24 @@ import { debug, getInput, info, setFailed } from '@actions/core'
 import { getOctokit } from '@actions/github'
 import glob from 'glob-promise'
 import {
-  DevelopBranches,
   getRef,
-  getRepository,
-  getSha,
-  getShaFromRef,
   getToken,
   isDevelopBranch,
   isReleaseBranch,
-  ReleaseBranches,
 } from '../../lib/github'
 import {
-  AnalysisResult,
-  filehashToGuiUrl,
-  guiUrlToFilehash,
-  VirusTotal,
-} from './virustotal'
-
-type CommitStatus = Awaited<
-  ReturnType<
-    ReturnType<typeof getOctokit>['rest']['repos']['listCommitStatusesForRef']
-  >
->['data'][number] & { sha: string }
+  CommitStatus,
+  getPendingVirusTotalStatuses,
+  writeStatus,
+} from './status'
+import { filehashToGuiUrl, guiUrlToFilehash, VirusTotal } from './virustotal'
 
 const ModeAnalyse = 'analyse'
 const ModeCheck = 'check'
 
 async function analyse(vt: VirusTotal, filePath: string) {
-  const result = await vt.scanFile(filePath)
+  const uploadUrl = await vt.getFileUploadURL()
+  const result = await vt.uploadFile(filePath, uploadUrl)
   info(`File "${filePath}" has been submitted for a scan`)
   info(`Analysis URL: ${filehashToGuiUrl(result.filehash)}`)
   return result
@@ -43,91 +33,6 @@ async function check(vt: VirusTotal, commitStatus: CommitStatus) {
     `Current status is "${result.status}" with "${result.flagged}" flags from security vendors`
   )
   return result
-}
-
-async function writeStatus(
-  octokit: ReturnType<typeof getOctokit>,
-  result: AnalysisResult,
-  sha?: string
-) {
-  await octokit.rest.repos.createCommitStatus({
-    owner: 'exivity',
-    repo: getRepository().component,
-    sha: sha ?? (await getSha()),
-    state:
-      result.status === 'pending'
-        ? 'pending'
-        : result.flagged === 0
-        ? 'success'
-        : 'failure',
-    context: `virustotal (${result.filename})`,
-    description:
-      result.status === 'completed'
-        ? result.flagged
-          ? `Detected as malicious or suspicious by ${result.flagged} security vendors`
-          : 'No security vendors flagged this file as malicious'
-        : undefined,
-    target_url: filehashToGuiUrl(result.filehash),
-  })
-  info('Written commit status')
-}
-
-async function getPendingVirusTotalStatuses(
-  octokit: ReturnType<typeof getOctokit>
-) {
-  const refs = [...ReleaseBranches, ...DevelopBranches]
-  const statuses: CommitStatus[] = []
-  for (const ref of refs) {
-    info(`Checking all statuses for ${ref}`)
-    try {
-      const component = getRepository().component
-      const sha = await getShaFromRef({
-        octokit,
-        component,
-        ref,
-        useFallback: false,
-      })
-      const { data } = await octokit.rest.repos.listCommitStatusesForRef({
-        owner: 'exivity',
-        repo: component,
-        ref: sha,
-      })
-      debug(`Total statuses: ${data.length}`)
-      // Results are in reverse chronological order, ignore any non-unique
-      // subsequent statuses
-      const uniqueStatuses = data.filter(
-        (status, i, arr) =>
-          arr.findIndex((s) => s.context === status.context) === i
-      )
-      debug(`Total unique statuses: ${uniqueStatuses.length}`)
-      if (!uniqueStatuses.length) {
-        info(`No pending virustotal statuses found`)
-      } else {
-        for (const status of uniqueStatuses) {
-          if (
-            status.context.startsWith('virustotal') &&
-            status.state === 'pending'
-          ) {
-            info(`Found pending virustotal status "${status.context}"`)
-            statuses.push({ ...status, sha })
-          }
-        }
-      }
-    } catch (error: unknown) {
-      if (error instanceof Error) {
-        if (
-          error.message.includes('Not Found') ||
-          error.message.includes('Branch not found')
-        ) {
-          info(`No commits found for branch ${ref}`)
-        } else {
-          throw error
-        }
-      }
-    }
-  }
-
-  return statuses
 }
 
 async function run() {
